+++ app/src/main/kotlin/com/pitchedapps/frost/contracts/ActivityContract.kt
@@ -1,0 +1,12 @@
+package com.pitchedapps.frost.contracts
+
+import io.reactivex.subjects.PublishSubject
+
+/**
+ * All the contracts for [MainActivity]
+ */
+interface ActivityContract : FileChooserActivityContract {
+    fun setTitle(titleRes: Int)
+    fun setTitle(title: String)
+    val fragmentSubject: PublishSubject<Int>
+}
