--- app/src/main/kotlin/com/pitchedapps/frost/contracts/WebContract.kt
@@ -1,8 +1,0 @@
-package com.pitchedapps.frost.contracts
-
-/**
- * Created by Allan Wang on 2017-07-04.
- *
- * Combination of all the core functions implemented by the Activity
- */
-interface ActivityWebContract : FileChooserActivityContract
