+++ app/src/main/kotlin/com/pitchedapps/frost/settings/Experimental.kt
@@ -40,7 +40,7 @@ fun SettingsActivity.getExperimentalPrefs(): KPrefAdapterBuilder.() -> Unit = {
     plainText(R.string.restart_frost) {
         descRes = R.string.restart_frost_desc
         onClick = {
-            setFrostResult(MainActivity.REQUEST_RESTART_APPLICATION)
+            setFrostResult(REQUEST_RESTART_APPLICATION)
             finish()
         }
     }
